+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add "Particle" effect for simple repeating particles
  - Add "Sound" effect for playing sounds or custom sounds, supporting volume and pitch-shifting
  - Add "probability" parameter for randomizing effects
+ - Add "orient" parameter to Atom effect
 
 # 6.3